+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix WhitePuffBall age not being reset when picked, Don't call super.use on ages less than 1
 - Fix WhitePuffBall growth being incorrect
 - Call GameEvent#Shear on successful FireCracker Leaves and Hydrangea Hedge Shears
+- Update WTHIT Plugin to new Format 
 
 # 1.1.2-Beta
 - Set getBucketItemStack in ManOWar.java to correct Item